add: tests and moved them to a separate test file
