+++ src/app/__tests__/game.test.jsx
@@ -5,8 +5,7 @@ import Wordle from '../../components/Wordle';
 import Grid from '../../components/Grid';
 import Row from '../../components/Row';
 import Tile from '../../components/Tile';
-import ToastMessage from '../../components/ToastMessage';
-import EndModal from '../../components/EndModal';
+
 import {
   formatGuess,
   addNewGuess,
@@ -393,83 +392,4 @@ test('tile no tiene la clase flip cuando no se le pasa bgColor', () => {
   render(<Tile letter="w" />);
   const tile = screen.getByTestId('tile');
   expect(tile).not.toHaveClass('flip');
-});
-
-test('despliega mensaje de exito y solucion cuando se gana', () => {
-  render(<EndModal isCorrect solution="hello" modalVisible />);
-
-  expect(screen.getByText('¡Adivinaste!')).toBeInTheDocument();
-
-  expect(screen.getByText('La palabra era HELLO.')).toBeInTheDocument();
-});
-
-test('ToastMessage despliega el mensaje correcto', () => {
-  render(<ToastMessage message="🔥 ¡Increíble!" />);
-  expect(screen.getByText('🔥 ¡Increíble!')).toBeInTheDocument();
-});
-
-test('se aplican animaciones flip y jump cuando las condiciones son verdaderas', () => {
-  const bgColor = '#3A3A3C';
-  const borderColor = '#3A3A3C';
-  const flipDelay = 0.3;
-  const jumpDelay = 0.2;
-
-  render(
-    <Tile
-      letter="A"
-      bgColor={bgColor}
-      borderColor={borderColor}
-      flipDelay={flipDelay}
-      shouldJump
-      jumpDelay={jumpDelay}
-    />,
-  );
-
-  const tile = screen.getByTestId('tile');
-  const animationStyle = tile.style.animation;
-
-  expect(animationStyle).toContain('flip');
-  expect(animationStyle).toContain('jump');
-  expect(animationStyle).toContain(`${flipDelay + jumpDelay}s`);
-});
-
-test('se setea isWinningRow correctamente', () => {
-  const guesses = [
-    [
-      { key: 'a', color: 'green' },
-      { key: 'b', color: 'green' },
-      { key: 'c', color: 'green' },
-      { key: 'd', color: 'green' },
-      { key: 'e', color: 'green' },
-    ],
-  ];
-
-  const currentGuess = '';
-  const turn = 1;
-  const isCorrect = true;
-
-  render(
-    <Grid
-      guesses={guesses}
-      currentGuess={currentGuess}
-      turn={turn}
-      isCorrect={isCorrect}
-    />,
-  );
-
-  const rows = screen.getAllByTestId('row');
-  expect(rows.length).toBe(guesses.length);
-
-  rows.forEach((row, index) => {
-    const tiles = row.querySelectorAll('[data-testid="tile"]');
-    const hasJumpClass = Array.from(tiles).some((tile) =>
-      tile.classList.contains('jump'),
-    );
-
-    if (index === turn - 1) {
-      expect(hasJumpClass).toBe(true);
-    } else {
-      expect(hasJumpClass).toBe(false);
-    }
-  });
 });
